move write gmcm xml method to adapter class
