+++ nzshm_model/gmcm_logic_tree/logic_tree.py
@@ -1,24 +1,23 @@
 import json
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, List, Union, Type
 
 import dacite
 
-from nzshm_model.psha_adapter import NrmlDocument
+from nzshm_model.psha_adapter import PshaAdapterInterface
 
+# def _process_gmm_args(args: List[str]) -> Dict[str, Any]:
+#     def clean_string(string):
+#         return string.replace('"', '').replace("'", '').strip()
 
-def _process_gmm_args(args: List[str]) -> Dict[str, Any]:
-    def clean_string(string):
-        return string.replace('"', '').replace("'", '').strip()
+#     args_dict = dict()
+#     for arg in args:
+#         if '=' in arg:
+#             k, v = arg.split('=')
+#             args_dict[clean_string(k)] = clean_string(v)
 
-    args_dict = dict()
-    for arg in args:
-        if '=' in arg:
-            k, v = arg.split('=')
-            args_dict[clean_string(k)] = clean_string(v)
-
-    return args_dict
+#     return args_dict
 
 
 @dataclass
@@ -36,44 +35,44 @@ class BranchSet:
 
 @dataclass
 class GMCMLogicTree:
-    version: str
-    title: str
+    version: str = ''
+    title: str = ''
     branch_sets: List[BranchSet] = field(default_factory=list)
 
-    @staticmethod
-    def from_xml(xml_path: Union[Path, str]) -> 'GMCMLogicTree':
-        """
-        Build a GMCMLogicTree from an OpenQuake nrml gmcm logic tree file.
-        """
-        doc = NrmlDocument.from_xml_file(xml_path)
-        if len(doc.logic_trees) != 1:
-            raise ValueError("xml must have only 1 logic tree")
+    # @staticmethod
+    # def from_xml(xml_path: Union[Path, str]) -> 'GMCMLogicTree':
+    #     """
+    #     Build a GMCMLogicTree from an OpenQuake nrml gmcm logic tree file.
+    #     """
+    #     doc = NrmlDocument.from_xml_file(xml_path)
+    #     if len(doc.logic_trees) != 1:
+    #         raise ValueError("xml must have only 1 logic tree")
 
-        branch_sets = []
-        for branch_set in doc.logic_trees[0].branch_sets:
-            branches = []
-            for branch in branch_set.branches:
-                if len(branch.uncertainty_models) != 1:
-                    raise ValueError('gmpe branches must have only one uncertainty model')
-                gmpe_name = branch.uncertainty_models[0].gmpe_name.replace('[', '').replace(']', '')
-                branches.append(
-                    Branch(
-                        gsim_clsname=gmpe_name,
-                        gsim_args=_process_gmm_args(branch.uncertainty_models[0].arguments),
-                        weight=branch.uncertainty_weight,
-                    )
-                )
-            branch_sets.append(
-                BranchSet(
-                    tectonic_region_type=branch_set.applyToTectonicRegionType,
-                    branches=branches,
-                )
-            )
-        return GMCMLogicTree(
-            version='',
-            title=doc.logic_trees[0].logicTreeID,
-            branch_sets=branch_sets,
-        )
+    #     branch_sets = []
+    #     for branch_set in doc.logic_trees[0].branch_sets:
+    #         branches = []
+    #         for branch in branch_set.branches:
+    #             if len(branch.uncertainty_models) != 1:
+    #                 raise ValueError('gmpe branches must have only one uncertainty model')
+    #             gmpe_name = branch.uncertainty_models[0].gmpe_name.replace('[', '').replace(']', '')
+    #             branches.append(
+    #                 Branch(
+    #                     gsim_clsname=gmpe_name,
+    #                     gsim_args=_process_gmm_args(branch.uncertainty_models[0].arguments),
+    #                     weight=branch.uncertainty_weight,
+    #                 )
+    #             )
+    #         branch_sets.append(
+    #             BranchSet(
+    #                 tectonic_region_type=branch_set.applyToTectonicRegionType,
+    #                 branches=branches,
+    #             )
+    #         )
+    #     return GMCMLogicTree(
+    #         version='',
+    #         title=doc.logic_trees[0].logicTreeID,
+    #         branch_sets=branch_sets,
+    #     )
 
     @staticmethod
     def from_dict(data: Dict) -> 'GMCMLogicTree':
@@ -87,3 +86,6 @@ class GMCMLogicTree:
 
     def to_dict(self) -> Dict[str, Any]:
         return asdict(self)
+
+    def psha_adapter(self, provider: Type[PshaAdapterInterface], **kwargs):
+        return provider(self)
